Moved all metrics calculations into Recomendations class
We wanted to combine the printer for the metrics and pinned IRQs into a single table.
Combining the get_percentage and get_total_num_interrupts methods into the Recommendations
class made this simpler and reduced the number of classes and thus made things easier
to read and understand.  The metrics relate to the recommendation so it can make sense
to have them in the Recommendation class like they are now.